Resolve "Add cooldown to missiles & add prediction targeting"

Closes #85, #89, #78, #83, #88, #92, and #86

See merge request waldemarlehner-hrw/ws2021/spieleentwicklung/space-shooter!65

+++ Assets/Scripts/Ship/Weaponry/Trigger/BasicAutoFireWeaponTrigger.cs
@@ -24,7 +24,7 @@ namespace Ship.Weaponry.Trigger
 
         public WeaponTriggerState CurrentState { get; protected set; } = WeaponTriggerState.NotFiring;
         public float TimeBetweenShots => this.cfg.MinTimeBetweenShots / (this.ShotDelayUpgradeLevel+1);
-        public int ShotDelayUpgradeLevel { get; set; }
+        public float ShotDelayUpgradeLevel { get; set; }
 
         public void Update(float dTime)
         {
