+++ Assets/Scripts/Ship/Weaponry/Trigger/BasicManualFireWeaponTrigger.cs
@@ -34,7 +34,7 @@ namespace Ship.Weaponry.Trigger
         public event Action? WeaponFiredEvent;
         public WeaponTriggerState CurrentState { get; protected set; } = WeaponTriggerState.NotFiring;
         public float TimeBetweenShots => this.cfg.MinTimeBetweenShots / this.ShotDelayUpgradeLevel;
-        public int ShotDelayUpgradeLevel { get; set; }
+        public float ShotDelayUpgradeLevel { get; set; }
 
         public void Update(float dTime)
         {
